# Use poolstl to sort randstrobes in parallel (attempt two)