--- a/src/index.cpp
+++ b/src/index.cpp
@@ -11,6 +11,7 @@
 #include <cassert>
 #include <algorithm>
 #include "pdqsort/pdqsort.h"
+#include "poolstl/poolstl.hpp"
 #include <iostream>
 #include <thread>
 #include <atomic>
@@ -138,7 +139,7 @@ int StrobemerIndex::pick_bits(size_t size) const {
     return std::clamp(static_cast<int>(log2(estimated_number_of_randstrobes)) - 1, 8, 31);
 }
 
-void StrobemerIndex::populate(float f, size_t n_threads) {
+void StrobemerIndex::populate(float f, unsigned n_threads) {
     Timer count_hash;
     auto randstrobe_counts = count_all_randstrobes(references, parameters, n_threads);
     stats.elapsed_counting_hashes = count_hash.duration();
@@ -164,8 +165,13 @@ void StrobemerIndex::populate(float f, size_t n_threads) {
 
     Timer sorting_timer;
     logger.debug() << "  Sorting ...\n";
-    // sort by hash values
-    pdqsort_branchless(randstrobes.begin(), randstrobes.end());
+    if (true) {
+        task_thread_pool::task_thread_pool pool{n_threads};
+        std::sort(poolstl::par.on(pool), randstrobes.begin(), randstrobes.end());
+    } else {
+        // sort by hash values
+        pdqsort_branchless(randstrobes.begin(), randstrobes.end());
+    }
     stats.elapsed_sorting_seeds = sorting_timer.duration();
 
     Timer hash_index_timer;