--- a/tests/compare-baseline.sh
+++ b/tests/compare-baseline.sh
@@ -13,8 +13,12 @@ set -euo pipefail
 python3 -c 'import pysam'
 
 ends="pe"
-while getopts "s" opt; do
+threads=4
+while getopts "st:" opt; do
   case "${opt}" in
+    t)
+      threads=$OPTARG
+      ;;
     s)
       ends=se  # single-end reads
       ;;
@@ -38,7 +42,7 @@ baseline_commit=$(< tests/baseline-commit.txt)
 baseline_bam=baseline/bam/${baseline_commit}.${ends}.bam
 baseline_binary=baseline/strobealign-${baseline_commit}
 cmake_options=-DCMAKE_BUILD_TYPE=RelWithDebInfo
-strobealign_options="-t 4"
+strobealign_options="-t ${threads}"
 
 # Generate the baseline BAM if necessary
 mkdir -p baseline/bam