We tried this before in #375, but one reason for closing the PR was that it was not possible to use a custom sort function (`pdqsort_branchless` in our case).

@alugowski has now added a `pluggable_sort` function, so we can try again.

Here are the sorting runtimes and memory usages for CHM13:

threads | only sorting | max mem
-|-|-
pdqsort_branchless (main) | 33 s | 14.3 GB
pluggable_sort, 1 thread  | 33 s | 14.3 GB
pluggable sort, 2 threads | 24 s | 16.8 GB
pluggable sort, 4 threads | 21 s | 16.8 GB
pluggable sort, 8 threads | 21 s | 16.8 GB

This is actually the first time I’m able to index CHM13 in under one minute (total indexing time) on my (home) machine, which is quite nice. However, memory usage still increases quite a bit.

@alugowski Is the extra memory usage expected?