@marcelm 2024-01-15
Hm, I guess the answer may be that [`std::inplace_merge` uses extra memory](https://stackoverflow.com/questions/11367674/inplace-merge-what-causes-a-complexity-of-nlogn-vs-n-1).