@marcelm 2024-01-15
Ah, [C++ inplace is not actually "in place"](https://stackoverflow.com/questions/4373307/is-it-possible-to-do-an-inplace-merge-without-temporary-storage/4375732#comment4762870_4373307) (as I would have understood it).