@marcelm 2024-01-16
> You're right, I also didn't realize how much memory `inplace_merge` might actually allocate. Is that too much for your application?

It probably is. We’ve been successfully reducing memory usage of the program over the last versions and this would be a step backwards. The ~12 seconds that we save here are just spent once at the startup of the program, and then for typical use cases it runs for at least a couple of minutes or tens of minutes, so it’s just a small fraction of the total. Increased memory usage on the other hand would mean that some type of data can not be analyzed on a machine with a given amount of RAM.

The program creates an index data structure for a genome. The 14 GB vs 16 GB I provided above were memory usage for a particular human genome (CHM13), which we use for testing. We want strobealign to work with much bigger genomes where the absolute increase would be much higher.

> `pluggable_sort` also lets you replace inplace_merge with another implementation. I see papers about algorithms that use less temp space, but don't see ready to try code. Just for fun I tried a no-op to see which step dominates, and in my naive benchmark the merges account for over half of the runtime.
> 
> The apparent limit around 21s is likely memory. This can be confirmed by testing with an artificially computationally expensive comparator. Basically a `std::less` with a busy loop. The runtimes would obviously be longer, but you'd see scaling as memory wouldn't be the bottleneck anymore.

Interesting. I probably don’t have time to do this experiment, but will keep it in mind.

I’ll leave the PR open until the owner of the repo @ksahlin has had time to comment (I know he’s busy at the moment).