@alugowski 2024-01-18
I went down an `inplace_merge` rabbit hole, and found something that seems usable. [This SO post](https://stackoverflow.com/a/4375732) references [an implementation](http://keithschwarz.com/interesting/code/?dir=inplace-merge) of the "Practical In-Place Merging" paper.  That code only supports merging ranges of equal sizes, but `pluggable_sort` emits as close to even sizes as possible and a naive adapter seems to work well. On my simple multigig sort benchmark pdqsort+this is faster than the plain poolSTL sort, and if the dataset is large enough to cause memory swapping then it's the outright fastest method (beating out pdqsort+std::inplace_merge). 

```c++
#include "thirdparty/InplaceMerge.hh"

/**
 * Like `std::inplace_merge` but uses a buffer-free algorithm described by Huang
 * and Langston in "Practical In-Place Merging", Communications of the ACM, 1988, http://dx.doi.org/10.1145/42392.42403
 * Generalizes the implementation by Keith Schwarz: http://keithschwarz.com/interesting/code/?dir=inplace-merge
 *
 * Schwarz's implementation only supports merging two ranges of equal size. This adapter also handles cases where the
 * ranges are slightly different sizes, as happens when used as the merge step for a general-purpose sort.
 * Drastically different sizes effectively fallback to std::inplace_merge, which is not buffer-free.
 */
template<class RandIt, class Compare>
void adapted_pipm_inplace_merge(RandIt first, RandIt mid, RandIt last, Compare comp) {
    auto left_size = std::distance(first, mid);
    auto right_size = std::distance(mid, last);

    if (left_size == right_size) {
        InplaceMerge(first, last, comp);
    } else if (left_size < right_size) {
        auto extra = right_size - left_size;
        InplaceMerge(first, last - extra, comp);
        std::inplace_merge(first, last - extra, last, comp);
    } else {
        // left_size > right_size
        auto extra = left_size - right_size;
        InplaceMerge(first + extra, last, comp);
        std::inplace_merge(first, first + extra, last, comp);
    }
}
```
Again, original `InplaceMerge.hh` is here: http://keithschwarz.com/interesting/code/?dir=inplace-merge

Usage: just add as last argument to `pluggable_sort`:
```c++
poolstl::pluggable_sort(poolstl::par.on(pool), randstrobes.begin(), randstrobes.end(), pdqsort_branchless, adapted_pipm_inplace_merge);
```

I also tried the bufferless inplace merges from several libraries like wikisort, libc++ libstdc++, and cppsort's symmerge implementation. All of those are significantly slower than the above; it seems they have the fallback but don't expect anyone to need it.

Also note that supposedly `std::inplace_merge` only uses the additional buffer if there's memory for it. There is the slower non-buffer fallback, so you *should* be able to use on it even for larger datasets, as `std::inplace_merge` would fallback to the bufferless version if there isn't enough memory available. I agree with you though, I'd also prefer to see the measured memory usage be what I expect and not depend on that fallback behavior.