@marcelm 2024-01-18
Cool, thanks for this! I had also found that `InplaceMerge.hh` file during my searching, but would not have been able to adapt it for use with pluggable_sort.

I tested this in commit 699cbe1. So the good news is that indeed memory usage remains unchanged no matter how many threads are used. The bad news is that it seems to become *slower* with more threads:

threads | wall-clock time (only sorting)
-|-
1 | 30.5 s
2 | 29.7 s
4 | 33.2 s
8 | 39.0 s

I’m not sure what is going on (given that you see a speedup).

> Also note that supposedly `std::inplace_merge` only uses the additional buffer if there's memory for it. There is the slower non-buffer fallback, so you _should_ be able to use on it even for larger datasets, as `std::inplace_merge` would fallback to the bufferless version if there isn't enough memory available. I agree with you though, I'd also prefer to see the measured memory usage be what I expect and not depend on that fallback behavior.

Yes, the "if there’s memory for it" part is way too vague. I believe this cannot be reliably detected.