@marcelm 2024-02-14
Thanks once again for your advice! We actually had `std::tie` in the comparison function, but I removed it in 26b1992c0899b367099684a326cf927d8e0f52f2 to speed up sorting. I think you have a good point that unstable sorting is not necessarily deterministic and that it would actually be good for reproducibility to compare more than just the hash. But more important is of course that we could get parallel sorting *and* be reproducible.

The advice of using a branchless comparison is great. I tried your `branchless_compare`, and it works better than `std::tie`, but single-threaded sorting time still increased quite a bit from 32.4 to 42.4 seconds. I wasn’t quite ready to accept that, so I tried a version using 128 bit ints:
```c++
    bool operator< (const RefRandstrobe& other) const {
        __uint128_t lhs = (static_cast<__uint128_t>(hash) << 64) | position;
        __uint128_t rhs = (static_cast<__uint128_t>(other.hash) << 64) | other.position;
        return lhs < rhs;
    }
```
and this increases sorting time to just 35 seconds, which is quite acceptable given that we’re going to be sorting with multiple threads anyway most of the time.

Here’s the disassembly of the three comparison functions:
https://godbolt.org/z/TPYvT5ocK
`std::tie` is unfortunately not branchless, which of course explains its slowness.

I’ll read through that article you linked. We may be able to apply the techniques to some other code that we have in inner loops.

## Runtimes for sorting with one thread

Comparison|time
-|-
By hash only (what we had before) | 32 s
By hash and position using `std::tie` | 45 s
By hash and position using `branchless_compare` | 42 s
By hash and position using a `__uint128_t` | 35 s

## Final runtimes

This was measured on an AMD CPU with four cores and hyperthreading
    
threads | sorting time | index creation time | previous index creation time
-|-|-|-
1 | 35 s | 154 s | 148 s
2 | 25 s | 95 s | 99 s
4 | 16 s | 60 s | 75 s
8 | 15 s | 48 s | 64 s

Thanks again for reaching out! I’m going to merge this as soon as CI passes.